+++ src/App.css
@@ -69,3 +69,14 @@ button {
   all: unset;
   cursor: pointer;
 }
+
+.a11y-hidden {
+    clip: rect(1px, 1px, 1px, 1px);
+    clip-path: inset(50%);
+    width: 1px;
+    height: 1px;
+    margin: -1px;
+    overflow: hidden;
+    padding: 0;
+    position: absolute;
+}
