+++ src/components/Intro/Intro.scss
@@ -4,7 +4,7 @@
   background-color: var(--intro-background);
   color: var(--color);
 
-  h1 {
+  h2 {
     font-family: "Poppins", sans-serif;
     font-style: normal;
     padding-top: 40px;
