+++ src/components/Project/Project.scss
@@ -41,7 +41,7 @@
     opacity: 1;
   }
 
-  h1 {
+  .title {
     font-family: "Poppins", sans-serif;
     font-style: normal;
     font-weight: 700;
